requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
python-dotenv>=1.1.1
pydantic>=2.11.7
requests>=2.32.4
streamlit>=1.36.0
langchain>=0.2.11 # Versão mais recente e alinhada
langchain-core>=0.2.16 # Versão alinhada com langchain
langchain-community>=0.2.12 # Versão alinhada
langchain-chroma>=0.2.4 # Versão alinhada
langchain-google-genai>=2.1.6 # Versão alinhada (já estava ok)
langchain-huggingface>=0.3.0 # Versão alinhada
langchain-ollama>=0.3.3 # Versão alinhada
langchain-text-splitters>=0.3.8 # Versão alinhada
pypdf>=4.2.0
python-docx>=1.1.0 # Para .docx, se ainda usar o ingest.py original
docopt>=0.6.2
torch>=2.3.0 # Para embeddings (InstructorEmbeddings)
transformers>=4.42.0 # Para embeddings (InstructorEmbeddings)
sentence-transformers>=3.0.0 # Para embeddings (InstructorEmbeddings)
onnxruntime>=1.17.1 # Para embeddings (InstructorEmbeddings)
accelerate>=0.31.0 # Para transformers
bitsandbytes>=0.43.1 # Para transformers